pipico/send_env_data_to_mtd/plot/requirements.txt
pyserial==3.5
pyqtgraph==0.14.0
PyQt5==5.15.11
numpy